pyproject.toml
[project]
name = "fastapi-taskmanager"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.10.1",
    "pymongo>=4.11.1",
    "python-jose[cryptography]>=3.4.0",
    "uvicorn>=0.34.0",
]
